requirements.txt
openai
faiss-cpu
numpy
scikit-learn
python-dotenv